+++ tests/inputs/kmod/create_reference_data.py
@@ -5,11 +5,14 @@ Run as module to avoid import errors.
 python -m tests.inputs.kmod.create_reference_data
 ```  
 """
-import tfs
 from pathlib import Path
+
+import tfs
+
 from omc3.model_creator import create_instance_and_model
 from tests.unit.test_kmod_averaging import update_reference_files as update_averages
-from tests.unit.test_kmod_import import update_reference_files as update_imports, _get_model_path 
+from tests.unit.test_kmod_import import get_model_path
+from tests.unit.test_kmod_import import update_reference_files as update_imports
 from tests.unit.test_kmod_lumi_imbalance import update_reference_files as update_lumi_imbalances
 
 THIS_DIR = Path(__file__).parent
@@ -35,7 +38,7 @@ def create_model(beam: int):
     # Compress ---
     df = tfs.read(THIS_DIR / f"b{beam}_model" / "twiss_elements.dat", index="NAME")
     df = df.loc[df.index.str.match("^(M|B|IP\d$)"), ["S", "BETX", "BETY"]]
-    tfs.write(_get_model_path(beam), df, save_index="NAME")
+    tfs.write(get_model_path(beam), df, save_index="NAME")
 
 
 if __name__ == "__main__":