+++ tests/unit/test_kmod_averaging.py
@@ -4,6 +4,7 @@ import pandas.testing as pdt
 import pytest
 import tfs
 
+from omc3.optics_measurements.constants import BETA, NAME
 from omc3.scripts.kmod_average import (
     AVERAGED_BETASTAR_FILENAME,
     AVERAGED_BPM_FILENAME,
@@ -20,20 +21,20 @@ REFERENCE_DIR = KMOD_INPUT_DIR / "references"
 @pytest.mark.parametrize("ip", [1, 5], ids=ids_str("ip{}"))
 @pytest.mark.parametrize("n_files", [1, 2], ids=ids_str("{}files"))
 def test_kmod_averaging(tmp_path, ip, n_files):
-    beta = 0.22
-    meas_paths = [_get_measurement_dir(ip, i+1) for i in range(n_files)]
-    ref_output_dir = _get_reference_dir(ip, n_files)
+    beta = get_betastar_model(beam=1, ip=ip)
+    meas_paths = [get_measurement_dir(ip, i+1) for i in range(n_files)]
+    ref_output_dir = get_reference_dir(ip, n_files)
 
     average_kmod_results(
         meas_paths=meas_paths, 
         output_dir=tmp_path,
         ip=ip, 
-        betastar=[beta],
+        betastar=beta,
         plot=True
      )
-    _assert_correct_files_are_present(tmp_path, ip, beta)
+    _assert_correct_files_are_present(tmp_path, ip, beta[0])
 
-    for out_name in _get_all_tfs_filenames(ip, beta):
+    for out_name in get_all_tfs_filenames(ip, beta[0]):
         out_file = tfs.read(tmp_path / out_name)
         ref_file = tfs.read(ref_output_dir / out_name)
         pdt.assert_frame_equal(out_file, ref_file, check_like=True)
@@ -41,20 +42,29 @@ def test_kmod_averaging(tmp_path, ip, n_files):
 
 def _assert_correct_files_are_present(outputdir: Path, ip: int, beta: float) -> None:
     """Simply checks the expected converted files are present in the outputdir"""
-    all_files = _get_all_tfs_filenames(ip, beta) + [f"ip{ip}_{PARAM_WAIST}.pdf", f"ip{ip}_{PARAM_BETA}.pdf"]
+    all_files = get_all_tfs_filenames(ip, beta) + [f"ip{ip}_{PARAM_WAIST}.pdf", f"ip{ip}_{PARAM_BETA}.pdf"]
     for file_name in all_files:
         assert (outputdir / file_name).is_file()
 
 
-def _get_measurement_dir(ip: int, i_meas: int) -> Path:
+def get_measurement_dir(ip: int, i_meas: int) -> Path:
     return KMOD_INPUT_DIR / f"ip{ip}_meas{i_meas}"
 
 
-def _get_reference_dir(ip: int, n_files: int) -> Path:
+def get_reference_dir(ip: int, n_files: int) -> Path:
     return REFERENCE_DIR / f"ip{ip}_averaged_{n_files}files"
 
 
-def _get_all_tfs_filenames(ip: int, beta:float) -> list[str]:
+def get_model_path(beam: int) -> Path:
+    return KMOD_INPUT_DIR / f"b{beam}_twiss_22cm.dat"
+
+
+def get_betastar_model(beam: int, ip: int) -> Path:
+    model = tfs.read(get_model_path(beam), index=NAME)
+    return model.loc[f"IP{ip}", [f"{BETA}Y", f"{BETA}Y"]].tolist()
+
+
+def get_all_tfs_filenames(ip: int, beta:float) -> list[str]:
     return [
         f"{AVERAGED_BPM_FILENAME.format(betastar_x=beta, betastar_y=beta, ip=ip, beam=1)}{EXT}",
         f"{AVERAGED_BPM_FILENAME.format(betastar_x=beta, betastar_y=beta, ip=ip, beam=2)}{EXT}",
@@ -67,16 +77,16 @@ def _get_all_tfs_filenames(ip: int, beta:float) -> list[str]:
 def update_reference_files():
     """ Helper function to update the reference files. """
     REFERENCE_DIR.mkdir(exist_ok=True, parents=True)
-    beta = 0.22
     for ip in (1, 5):
+        beta = get_betastar_model(beam=1, ip=ip)
         for n_files in (1, 2):
-            meas_paths = [_get_measurement_dir(ip, i+1) for i in range(n_files)]
-            output_dir = _get_reference_dir(ip, n_files)
+            meas_paths = [get_measurement_dir(ip, i+1) for i in range(n_files)]
+            output_dir = get_reference_dir(ip, n_files)
             average_kmod_results(
                 meas_paths=meas_paths, 
                 output_dir=output_dir,
                 ip=ip, 
-                betastar=[beta],
+                betastar=beta,
                 plot=False
             )
             for ini_file in output_dir.glob("*.ini"):