+++ tests/unit/test_kmod_import.py
@@ -4,7 +4,6 @@ import pandas.testing as pdt
 import pytest
 import tfs
 
-from omc3.scripts.kmod_import import import_kmod_data
 from omc3.optics_measurements.constants import (
     AVERAGED_BETASTAR_FILENAME,
     AVERAGED_BPM_FILENAME,
@@ -12,21 +11,25 @@ from omc3.optics_measurements.constants import (
     BETA_STAR_FILENAME,
     EXT,
 )
-from tests.conftest import INPUTS
-from tests.unit.test_kmod_averaging import _get_reference_dir as get_averages_dir
+from omc3.scripts.kmod_import import import_kmod_data
+from tests.unit.test_kmod_averaging import (
+    KMOD_INPUT_DIR,
+    get_betastar_model,
+    get_model_path,
+)
+from tests.unit.test_kmod_averaging import get_reference_dir as get_averages_dir
 
-INPUT_DIR_KMOD = INPUTS / "kmod"
-REFERENCE_DIR = INPUT_DIR_KMOD / "references"
+REFERENCE_DIR = KMOD_INPUT_DIR / "references"
 IP1_RESULTS_OUTPUTS = REFERENCE_DIR / "ip1_averaged_2files"
 IP5_RESULTS_OUTPUTS = REFERENCE_DIR / "ip5_averaged_2files"
 
-B1_RESULTS_OUTPUTS = INPUT_DIR_KMOD / "b1_imported"
-B2_RESULTS_OUTPUTS = INPUT_DIR_KMOD / "b2_imported"
+B1_RESULTS_OUTPUTS = KMOD_INPUT_DIR / "b1_imported"
+B2_RESULTS_OUTPUTS = KMOD_INPUT_DIR / "b2_imported"
 
 @pytest.mark.basic
 @pytest.mark.parametrize('beam', [1, 2])
 def test_kmod_import_averaged_folder_beam(tmp_path, beam):
-    model = _get_model_path(beam)
+    model = get_model_path(beam)
 
     path_bpm_ip1 = get_averages_dir(ip=1, n_files=2)
     path_bpm_ip5 = get_averages_dir(ip=5, n_files=2) 
@@ -53,8 +56,8 @@ def test_kmod_import_averaged_folder_beam(tmp_path, beam):
 @pytest.mark.parametrize('files', ["bpm", "betastar", "bpm-betastar"])
 @pytest.mark.parametrize('read', [True, False])
 def test_kmod_import_files_beam(tmp_path, beam, files, read):
-    model = _get_model_path(beam)
-    beta = 0.22
+    model = get_model_path(beam)
+    beta = get_betastar_model(beam, ip=1)[0]
 
     paths = []
     if "bpm" in files:
@@ -99,10 +102,6 @@ def _assert_correct_files_are_present(outputdir: Path, which: str = "bpm-betasta
             assert (outputdir / f"{BETA_STAR_FILENAME}{plane}{EXT}").is_file()
 
 
-def _get_model_path(beam: int) -> Path:
-    return INPUT_DIR_KMOD / f"b{beam}_twiss_22cm.dat"
-
-
 def _get_bpm_input_path(beam: int, ip: int, beta: float) -> Path:
     return get_averages_dir(ip=ip, n_files=2) / f"{AVERAGED_BPM_FILENAME.format(betastar_x=beta, betastar_y=beta, ip=ip, beam=beam)}{EXT}"
 
@@ -124,18 +123,13 @@ def _get_betastar_referece_path(beam: int, plane: str) -> Path:
 def update_reference_files():
     """ Helper function to update the reference files. """
     REFERENCE_DIR.mkdir(exist_ok=True, parents=True)
-    beta = 0.22
     for beam in (1, 2):
         output_path = REFERENCE_DIR / f"b{beam}_imported"
         output_path.mkdir(exist_ok=True, parents=True)
-        model = _get_model_path(beam)
+        model = get_model_path(beam)
         path_ip1 = get_averages_dir(ip=1, n_files=2)
         path_ip5 = get_averages_dir(ip=5, n_files=2)
 
-        import_kmod_data(meas_paths=[path_ip1, path_ip5], model=model, output_dir=output_path, beam=beam)
+        import_kmod_data(measurements=[path_ip1, path_ip5], model=model, output_dir=output_path, beam=beam)
         for ini_file in output_path.glob("*.ini"):
             ini_file.unlink()
-
-
-if __name__ == "__main__":
-    update_reference_files()